psm: move get_power to ir_solver

+++ src/psm/src/ir_solver.h
@@ -101,7 +101,7 @@ class IRSolver
   //! Function to solve for IR drop
   void solveIR();
   //! Function to get the power value from OpenSTA
-  std::vector<std::pair<odb::dbInst*, double>> getPower();
+  std::vector<std::pair<odb::dbInst*, float>> getPower();
 
   bool getConnectionTest() const;
   int getMinimumResolution() const;
